This is the health check page for shared load balancing
To set up load balancing we will to have you put a health check page that our load balancers can use to verify the functionality of the site(s). The load balancer checks the MD5 checksum of the test page, if this changes (IE, if the contents of the page changes) the load balancer will remove the server from the load-balancing group.  For this reason, if you ever change the content of the health check page, please let us know so we can update the load balancer config file.

We recommend a separate dummy page for this that will not change so long as the test is working correctly. That way, if you update the home page, it will not kick the site out of the load balancing group. (If the contents of the page change at all the load balancers will see that as a failure.)

git-svn-id: http://svn2.assembla.com/svn/halotrac/trunk@228 04a73145-ec33-0410-b10b-c8dba00f0e28