refs #1523: new SafetyCareClient, heartbeat task, safetycare alert on critical device event
Needs the proper alert type numbers in event.event_type_numeric, and the SafetyCareClient to be changed to point to SafetyCare's servers before it goes live (it points to localhost for testing using the safetycare_test_listener right now, which just listens for incoming connections and prints out the first line of the response with a timestamp, printing an ACK if it sees something that isn't a @).

Added the 1m timeout to environment.rb to keep parity with existing tasks.